% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BISC-package.R
\docType{package}
\name{BISC-package}
\alias{BISC-package}
\alias{BISC}
\title{The 'BISC' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.11. https://mc-stan.org
}
