% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BISC_estimate.R
\name{BISC_estimate}
\alias{BISC_estimate}
\title{Transcriptional bursting estimation with Stan}
\usage{
BISC_estimate(data, model = "PB-trend", iter = 4000)
}
\arguments{
\item{data}{read count input data matrix (row: genes; column: cells).}

\item{model}{estimation model, defaults to "PB-trend". BISC includes three model options: PB model, PB-trend model and ZIPB-trend model}

\item{iter}{number of MCMC interations, defaults to 4000}
}
\value{
This function outputs a list object contains: 1.A data frame contains four columns: a. Gene name; b. Kon; c. koff; d.s;
2.MCMC outputs which could be used for differential bursting analysis.
}
\description{
This function estiamtes the transcriptional bursting parameters kon, koff and s from
the input gene expression read count data.
}
