% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_lm.R
\name{checks}
\alias{checks}
\title{Helper function for checking arguments}
\usage{
checks(
  x,
  one_element = TRUE,
  atomic = TRUE,
  numeric = TRUE,
  positive = TRUE,
  int = TRUE
)
}
\arguments{
\item{x}{The input argument you want to check.}
}
\value{
void, or throws an error message
}
\description{
Checks if an input argument is e.g. positive.
}
\examples{
  checks(regulator_mean, one_element=TRUE, atomic=TRUE, numeric=TRUE, positive=TRUE, int=FALSE)
}
