% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BISC_differential_bursting.R
\name{DB}
\alias{DB}
\title{Differential bursting analysis}
\usage{
DB(data1, data2, frequency, size, log2)
}
\arguments{
\item{data1}{output from BISC_estimate() of treatment condition.}

\item{data2}{output from BISC_estimate() of control condition.}

\item{frequency}{testing the H0: kon=frequency.}

\item{size}{testing the H0: size=size.}

\item{log2}{compute difference of bursting estimates in log2 scale: T or F}
}
\value{
This function outputs a list object contains testing results for three parameters (i.e. kon, koff and s).
}
\description{
This function perfoms the differential bursting analysis of bursting frequency(i.e. kon) and bursting size(s/koff)
between two study conditions (e.g. Tumor vs. Normal) using
MCMC output from BISC_estimate().
}
