% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_lm.R
\name{generate_data_lm}
\alias{generate_data_lm}
\title{Dummy data generation for biclust with lm}
\usage{
generate_data_lm(
  n_cell_clusters = 3,
  n_target_gene_type = 5,
  n_regulator_gene_type = 20,
  n_cells = c(1000, 5000, 10000),
  regulator_means = c(1, 2, 3),
  regulator_standard_deviations = c(0.1, 0.2, 0.3),
  coefficients_standard_deviation = 100,
  target_gene_type_standard_deviation = 3
)
}
\arguments{
\item{n_cell_clusters}{The number of cell clusters.}

\item{n_target_gene_type}{The number of target gene types. We have x named target genes that have one expression per cell.}

\item{n_regulator_gene_type}{The number of regulator gene types. We have x named regulator genes that have one expression per cell.}

\item{n_cells}{c(1000,5000,10000), The number of cells in each cell cluster given as a vector.}

\item{regulator_means}{c(1,2,3), Regulator gene expression mean in each cell cluster.}

\item{regulator_standard_deviations}{= c(0.1,0.2,0.3),  Regulator sd for expression in each cell cluster.}

\item{coefficients_standard_deviation}{= 100, 'The betas/slopes'. One per target gene. Instead of providing mean and standard deviation for each target gene, we provide the standard deviation from which these will be generated. Mean will be 0.}

\item{target_gene_type_standard_deviation}{= 3, The error, the sd of the expression of the target genes.}
}
\value{
dat
}
\description{
Generates dummy data that work with biclust with lm
For now we don't generate sparse data with regulators and targets set to 0. Also known as, we don't generate target gene type clusters.
That's for later when we want to play with e.g. lasso.
We assume:
ERROR: The error variance is the same across all cell clusters. The formal assumption for (many x -> many y linear regression) is:
  The errors are usually assumed to be uncorrelated across measurements, and follow a multivariate normal distribution. If the errors do not follow a multivariate normal distribution, generalized linear models may be used to relax assumptions about Y and U.
BETAS/COEFFICIENTS' MEANS AND SD:
  Close to zero mean and high variance, one different mean and sd per target gene (same across all regulator genes). This arranges it so that we have one slope per cell cluster.
REGULATOR MEANS AND SD:
  One mean and sd per cell cluster. It might be easier just to keep it the same for all cell clusters.
TODO: Go through the theory and assumptions of the model: https://en.wikipedia.org/wiki/Multivariate_normal_distribution
}
\examples{
  dat <- generate_data_lm(n_cell_clusters = 3,
                          n_target_gene_type = 5,
                          n_regulator_gene_type = 20,
                          n_cells = c(1000,5000,10000),
                          regulator_means = c(1,2,3),
                          regulator_standard_deviations = c(0.1,0.2,0.3),
                          coefficients_standard_deviation = 100,
                          target_gene_type_standard_deviation = 3,
                          )
}
