% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_data_for_scregclust.R
\name{generate_dummy_data_for_scregclust}
\alias{generate_dummy_data_for_scregclust}
\title{Dummy data generation for scRegClust}
\usage{
generate_dummy_data_for_scregclust(
  n_target_genes = 50,
  n_regulator_genes = 30,
  n_cells = 1000,
  n_target_gene_clusters = 3,
  regulator_mean = 1,
  coefficient_mean = c(30, 50, 100)
)
}
\arguments{
\item{n_target_genes}{The number of target genes}

\item{n_regulator_genes}{The number of regulator genes}

\item{n_cells}{The number of cells}

\item{n_target_gene_clusters}{The number of target gene clusters}

\item{regulator_mean}{Mean expression of regulator genes}

\item{coefficient_mean}{Mean coefficients in true model, length n_target_gene_clusters}
}
\value{
A list with
  \item{Z_t}{a}
  \item{Z_r}{a}
  \item{Pi}{n_target_gene_clusters x n_target_genes. True cell cluster allocation.}
  \item{R}{a}
  \item{S}{a}
  \item{B}{a}
}
\description{
Generates dummy data that work with scRegClust.
Rows are cluster index. Cols are target gene index.
}
\examples{
res <- generate_dummy_data_for_scregclust(50, 30, 1000, 3, 1, c(1,20,100));
}
