% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomise_cluster_labels.R
\name{randomise_cluster_labels}
\alias{randomise_cluster_labels}
\title{Randomise cluster labels}
\usage{
randomise_cluster_labels(cluster_labels, fraction_randomised = 0.05)
}
\arguments{
\item{cluster_labels}{Vector of cluster labels, integers, 1 to x, not skipping any numbers.}

\item{fraction_randomised}{The fraction of each cluster you want reassigned.}
}
\value{
A vector with randomized cluster labels
}
\description{
Take a vector of integer cluster labels from 1 to x, not skipping any numbers.
Assign a fraction fraction_randomised of them a different cluster label.
}
\examples{
cluster_labels <- rep(c(1,2,3,4), 10)
res <- randomise_cluster_labels(cluster_labels = cluster_labels,
                                fraction_randomised  = 0.5)
print(cluster_labels)
print(res)
}
