% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclust.R
\name{biclust}
\alias{biclust}
\title{Biclust}
\usage{
biclust(
  dat = dat,
  max_iter = 50,
  initial_clustering,
  n_target_genes,
  n_regulator_genes,
  n_total_cells,
  n_cell_clusters,
  ind_targetgenes,
  ind_reggenes,
  output_path,
  penalization_lambda = 0.5
)
}
\value{
Nothing yet, maybe cluster labels
}
\description{
Cluster cells based with Classification EM based on lm with penality
}
\examples{
NULL
}
