% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_data_for_scregclust.R
\name{generate_dummy_data_for_scregclust}
\alias{generate_dummy_data_for_scregclust}
\title{Dummy data generation for scRegClust}
\usage{
generate_dummy_data_for_scregclust(
  n_target_genes = 10,
  n_regulator_genes = 5,
  n_cells = 1000,
  n_target_gene_clusters = 3,
  regulator_mean = 1,
  coefficient_mean = c(1, 20, 100)
)
}
\arguments{
\item{n_target_genes}{The number of target genes}

\item{n_regulator_genes}{The number of regulator genes}

\item{n_cells}{The number of cells}

\item{n_target_gene_clusters}{The number of target gene clusters}

\item{regulator_mean}{Mean expression of regulator genes}

\item{coefficient_mean}{Mean coefficients in true model, length n_target_gene_clusters}
}
\value{
list(Z_t, Z_r, Pi, R, S, B)
}
\description{
Generates dummy data that work with scRegClust
}
\examples{
res <- generate_dummy_data_for_scregclust(50, 30, 1000, 3, 1, c(1,20,100));
}
